Updated code
Made countdown, difficulty select, draw function, state machine for the game, and a couple of LCD screen showings.

User can press any button that isn't the L-PB or R-PB to start. It will ask them to press again to actually start the game. If they instead pressed the L or R PB, it will return to difficulty select screen