Added a switch case statement for the gameState, added new LCD code (draw), made some small changes in all files so the code is easier to read.
In resources.h, a new function that acts as a non-blocking delay has been added.

In the LCD files, a show_watch function has been added that displays the word "WATCH" with an arrow on top.

Removed PLAYING gamestate and replaced it with REPEAT

In MAIN, i added a generatePattern() function that generates a random pattern that the player must repeat (used dynamic arrays). Added new switch case for changing gameState. Moved the LCD code into a protothread function which draws every 200ms.