Reworked the whole program in simon_says.ino, deleted some defines in ESP12E_DFINES.h
Recoded the game and lowered the amount of buttons and lights to 2 (Not enough GPIOs on the ESP8266). The only interrupts remaining are LEFT_ISR and RIGHT_ISR corresponding to D3 and D4. To make up for the 2 less buttons, a simultaneous button combo was added and now acts as the "enter key" or "center button" during the game. Added volatile keyword to FLAGS as it was mainly edited during an ISR. Added a resetPB() function to facilitate as the FLAGS reset.

 - Not yet edited the generation of the pattern. It still randomizes G,R,B,Y instead of LEFT, RIGHT, CENTER

Rewrote the ESP12E_DFINES.h as some of its definitions conflicted with arduino's pin defines.