Fixed push button logic
There was a bug where hitting two buttons simultaneously sometimes triggered an individual press flag causing an unexpected error or "double clicking". Also there were open ends in the code meaning the code didn't account for all possibilities and sometimes behaved wrong.

To fix the issue I recoded all button controls in the ISR and the loop by "funnelling" the flow of a code to a constricted point. Basically, it checks for a button press every 200ms and activates the correct flag for each button without it doubling a press.