Improved user input, added new LCD picture, WATCHING foundation has been made.
resources.h:
Removed the function "is_Pushed" as it is redundant and only used in the old user input code.
Added two new functions LEDOFF and lights.

LCD:
Added new LCD showcase: show_repeat. This screen gets displayed during the REPEAT game state.

simon_says.ino:
Created the foundation code for WATCHING. The watch function displays the order in which the lights are to be repeated. Currently, it shows the whole pattern all at once. Later, this will be redone and it will show one level at a time similar to simon says.